requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
scikit_learn==1.0.2
tensorflow==2.8.0
